Moved IE9 metadata and Google search templating to theme.php
Since NNF does not provide its own search-engine, the search box is
entirely part of the theme. I have moved the search templating out of
the core and into theme.php, as well as the IE9 metadata as I will not
mandate that for all themes.

WARNING: If you have a custom skin, note that you will need to provide
the search engine HTTP/HTTPS template code in your theme.php. You can
copy greyscale's theme.php for this.

Also fixed in this commit:
* Custom logo had not been working for a long time, and was using
entirely the wrong path
* Error message when stickies not present in sub-forums