Starting to get this into shape
Stuff learnt:
1. Previous change was pointless -- adding the lock/unlock querystring
to the REQUEST_URI breaks the pretty URLs. Lock/unlock have to go
through central URL function instead. Signin will be difficult to
convert as such due to `prepareTemplate` not having knowledge of the
current page number yet
2. All URLs have to be canonicalised into absolute URLs to be able to
cope with the massive variety of inputs and outputs
3. Absolute URLs trigger Apache's auto-correction of files into folders,
preventing you from using threads with the same name as a sub-forum
4. Absolute URLs need to be converted back into relative after
canonicalisation in order to solve the above problem