Fixed #143: Italics appearing in the middle of URLs
An "_&_" appearing in the middle of a URL was triggering italics because the regex "\w" was seeing the semicolon after "&amp;" as a sufficient non-word character. Have changed the regex to use "\S" (not whitespace) instead which fixes the immediate problem.

However, this breaks combined bold and italic ("*_..._*"), but that wasn't working consistently anyway (e.g. "*_..._*"); having looked at this, it raises a few deficiencies of the italic / bold regex, which I will file in another bug.