Fixed #132: Improve transliteration / use iconv optionally
NNF now uses PHP5.4's transliteration if available, which can convert arabic, greek, hebrew, japanese and even chinese
scripts to ASCII. If PHP5.4's transliteration is not available, iconv is attempted, and if that is missing, an overall
fallback is provided that will handle most instances of Latin-based languages.

With thanks to bh8(dot)vn for suggestion & zuchto for raising the iconv issue.